+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
+
+# v1.1.7(2024-01-xx)
+* Renamed api-proxy.yml server info field `ip` to `host`
+* In api-proxy.yml assign server config to user
+
 # v1.1.6(2024-01-17)
-* watch filter are now regular expressions 
+* Watch filter are now regular expressions 
 * Fixed watch file not created problem
 * UI responds immediately to update request